+++ platform/esp-idf/platform.c
@@ -49,6 +49,7 @@ int PlatformWappLoad(const char *name, wapp_t *wapp) {
 static const vfs_driver_table_t esp_driver_table[] = {
     {"wifi", VfsWifiInit},
     {"ota", VfsOtaInit},
+    {"gpio", VfsGpioInit},
     {NULL, NULL},
 };